Support for featured image
For the same reason as suggesting support for custom excerpt is added, I would suggest support for featured images is added. I have featured images on most of my posts and when I switched to Crisp I "lost" those images. Adding support gives people the option to use it if they want. This featured image code is borrowed from the default Casper theme.